refactor(tool-harness): extract transport + executors into modules (#304)
First half of #304: split the 1873-line scripts/run_tool_harness.py into focused
modules with no behavior change.

- pr_reviewer/transport.py — safe_run, run_chat_request, run_chat_completion
  (curl-based model-call transport + the shared subprocess runner).
- pr_reviewer/tool_executors.py — the read-only executors (read_file, git_grep/
  log/blame, gh_api, web_fetch, web_search, run_command), execute_tool_request[s],
  and the path/host guards they need.

run_tool_harness.py re-imports the moved names so every existing call site and
test that references them via run_tool_harness keeps working; it still owns the
planner, run_native_loop, and main (1873 -> 1230 lines). Moved code is a
byte-identical AST-extracted copy (verified).

Three tests monkeypatched these functions on run_tool_harness; since
execute_tool_requests / run_chat_request now call the module-local name, the
patches are repointed at the module that defines them (transport / tool_executors).

The planner removal (run_chat_completion becomes dead) is the next PR.
Part of the 2.0 consolidation epic (#258).
