+++ scripts/run_tool_harness.py
@@ -27,32 +27,36 @@ if str(_PROJECT_ROOT) not in sys.path:
 
 from redact import mask_secrets  # noqa: E402
 
-# The gh_api allowlist (path regex, prefixes, deny substrings) now lives
-# entirely on the platform seam (pr_reviewer.platform), which owns both the
-# GitHub and Forgejo backends — a single source of truth so the two can't
-# drift (the kind of tiny mapping mistake that becomes a CVE). We import only
-# GH_DENY_SUBSTRINGS here, reused below by _resolve_workspace_path to block
-# the same sensitive segments in filesystem paths.
-from pr_reviewer.platform import GH_DENY_SUBSTRINGS  # noqa: E402
-
-
-SENSITIVE_PATH_RE = re.compile(
-    r"(^|/)(\.env(\.|$)|id_rsa(\.|$)|id_dsa(\.|$)|credentials(\.|$)|secret(s)?(\.|$)|.*\.pem$|.*\.key$)",
-    re.IGNORECASE,
+# Transport + read-only executors were split into dedicated modules (#304).
+# Re-imported here so call sites and tests that reference these names via
+# run_tool_harness keep working unchanged.
+from pr_reviewer.transport import (  # noqa: E402
+    run_chat_completion,
+    run_chat_request,
+    safe_run,
 )
-
-# The tool harness executes same-repo code, so command execution must not be
-# model-controlled shell text. Keep commands as named, argv-only definitions.
-# Additions here should be read-only and safe to run against untrusted PR input.
-ALLOWED_COMMANDS = {
-    "git_status_short": ["git", "status", "--short"],
-    "git_diff_stat": ["git", "diff", "--stat", "HEAD"],
-    "git_diff_name_only": ["git", "diff", "--name-only", "HEAD"],
-}
-
-
-def command_catalog_markdown():
-    return ", ".join(sorted(ALLOWED_COMMANDS))
+from pr_reviewer.tool_executors import (  # noqa: E402
+    ALLOWED_COMMANDS,
+    SENSITIVE_PATH_RE,
+    _opt_int,
+    _resolve_workspace_path,
+    allowlisted_host,
+    command_catalog_markdown,
+    execute_tool_request,
+    execute_tool_requests,
+    fetch_url,
+    gh_api,
+    git_blame,
+    git_grep,
+    git_log,
+    mask_and_truncate,
+    normalize_host,
+    read_file,
+    run_command,
+    truncate_text,
+    web_fetch,
+    web_search,
+)
 
 
 def normalize_repo_name(value):
@@ -80,16 +84,6 @@ def env_int(name, default_value, min_value):
 def env_int_bounded(name, default_value, min_value, max_value):
     value = env_int(name, default_value, min_value)
     return min(max_value, value)
-
-
-def _opt_int(value):
-    """Coerce an optional tool arg to int, tolerating model string/None forms."""
-    if value is None:
-        return None
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return None
 
 
 def _accumulate_usage(acc, response, api_format):
@@ -148,27 +142,6 @@ def normalize_api_format(value):
     return "openai"
 
 
-def normalize_host(host):
-    return (host or "").strip().lower()
-
-
-def allowlisted_host(host, allowlist):
-    candidate = normalize_host(host)
-    for item in allowlist:
-        if candidate == item:
-            return True
-    return False
-
-
-def truncate_text(text, max_bytes):
-    masked = mask_secrets(text)
-    raw = masked.encode("utf-8", errors="replace")
-    if len(raw) <= max_bytes:
-        return masked, False
-    clipped = raw[:max_bytes].decode("utf-8", errors="replace")
-    return clipped + "\n[truncated]", True
-
-
 def extract_json_object(text):
     data = text.strip()
     if data.startswith("```"):
@@ -199,445 +172,6 @@ def extract_json_object(text):
         parsed = parsed[0]
 
     return parsed
-
-
-def mask_and_truncate(text, max_bytes):
-    masked = mask_secrets(text)
-    raw = masked.encode("utf-8", errors="replace")
-    if len(raw) <= max_bytes:
-        return masked, False
-    clipped = raw[:max_bytes].decode("utf-8", errors="replace")
-    return clipped + "\n[truncated]", True
-
-
-def safe_run(args, timeout_sec):
-    """Run a command and capture stdout/stderr with a timeout."""
-    try:
-        return subprocess.run(
-            args,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=False,
-            timeout=timeout_sec,
-        )
-    except subprocess.TimeoutExpired as exc:
-        return {
-            "timeout": True,
-            "stdout": (exc.stdout or "") if isinstance(exc.stdout, str) else "",
-            "stderr": (exc.stderr or "") if isinstance(exc.stderr, str) else "",
-        }
-
-
-def run_chat_request(base_url, api_format, payload, api_key, timeout_sec):
-    """POST a wire-ready chat payload via curl and return the parsed JSON.
-
-    Transport for the native tool-calling loop (#203): the payload is built
-    by ``pr_reviewer.conversation.Conversation.to_request_payload``, so this
-    function owns only the endpoint choice, auth, and JSON decode.
-    """
-    if api_format == "anthropic":
-        endpoint = base_url.rstrip("/") + "/messages"
-    else:
-        endpoint = base_url.rstrip("/") + "/chat/completions"
-
-    curl_args = [
-        "curl",
-        "-q",
-        "-fsSL",
-        "--max-time",
-        str(timeout_sec),
-        endpoint,
-        "-H",
-        "Content-Type: application/json",
-    ]
-    if api_format == "anthropic":
-        curl_args.extend(["-H", f"anthropic-version: {os.getenv('ANTHROPIC_VERSION', '2023-06-01')}"])
-
-    # Streaming keeps bytes flowing so proxies with a short idle/read timeout
-    # (Cloudflare's 100s edge timer etc.) don't 524 a long thinking-model turn.
-    # --no-buffer flushes each SSE chunk; the body is reassembled below.
-    streaming = bool(payload.get("stream"))
-    if streaming:
-        curl_args.append("--no-buffer")
-        if api_format == "anthropic":
-            curl_args.extend(["-H", "Accept: text/event-stream"])
-
-    # The API key goes through a 0600 curl --config file rather than argv, so
-    # it never appears in /proc/<pid>/cmdline or `ps` output on shared runners.
-    auth_config_path = None
-    if api_key:
-        if api_format == "anthropic":
-            auth_header = f"x-api-key: {api_key}"
-        else:
-            auth_header = f"Authorization: Bearer {api_key}"
-        escaped = auth_header.replace("\\", "\\\\").replace('"', '\\"')
-        fd, auth_config_path = tempfile.mkstemp()
-        with os.fdopen(fd, "w", encoding="utf-8") as auth_file:
-            auth_file.write(f'header = "{escaped}"\n')
-        os.chmod(auth_config_path, 0o600)
-        curl_args.extend(["--config", auth_config_path])
-
-    with tempfile.NamedTemporaryFile("w", encoding="utf-8", delete=False) as payload_file:
-        json.dump(payload, payload_file)
-        payload_path = payload_file.name
-
-    try:
-        completed = safe_run(curl_args + ["--data", f"@{payload_path}"], timeout_sec + 5)
-    finally:
-        for cleanup_path in (payload_path, auth_config_path):
-            if cleanup_path is None:
-                continue
-            try:
-                os.unlink(cleanup_path)
-            except OSError:
-                pass
-
-    if isinstance(completed, dict) and completed.get("timeout"):
-        raise RuntimeError("planner model request timed out")
-    if completed.returncode != 0:
-        stderr = mask_secrets((completed.stderr or "").strip())
-        if len(stderr) > 500:
-            stderr = stderr[:500] + "...[truncated]"
-        raise RuntimeError(
-            f"planner model request failed with exit code {completed.returncode}"
-            + (f": {stderr}" if stderr else "")
-        )
-
-    if streaming:
-        # SSE deltas → the non-streaming response shape the loop parses. The
-        # reassembler also surfaces a JSON error body returned mid-"stream"
-        # (some servers reply 200 + an error object instead of events).
-        from pr_reviewer.sse_reassembler import reassemble_sse  # noqa: PLC0415
-
-        return reassemble_sse(completed.stdout, api_format)
-    return json.loads(completed.stdout)
-
-
-def run_chat_completion(
-    base_url,
-    api_format,
-    model,
-    api_key,
-    system_prompt,
-    user_prompt,
-    timeout_sec,
-    max_tokens,
-):
-    """Call an AI model via curl and return the response text."""
-    if api_format == "anthropic":
-        payload = {
-            "model": model,
-            "system": system_prompt,
-            "messages": [{"role": "user", "content": user_prompt}],
-            "temperature": 0.0,
-            "max_tokens": max_tokens,
-        }
-    else:
-        payload = {
-            "model": model,
-            "messages": [
-                {"role": "system", "content": system_prompt},
-                {"role": "user", "content": user_prompt},
-            ],
-            "temperature": 0.0,
-            "max_tokens": max_tokens,
-        }
-
-    parsed = run_chat_request(base_url, api_format, payload, api_key, timeout_sec)
-    if api_format == "anthropic" and isinstance(parsed.get("content"), list):
-        parts = []
-        for item in parsed.get("content", []):
-            if isinstance(item, dict) and item.get("type") == "text":
-                text = item.get("text")
-                if isinstance(text, str):
-                    parts.append(text)
-        return "".join(parts)
-    return parsed.get("choices", [{}])[0].get("message", {}).get("content", "")
-
-
-def fetch_url(url, allowed_hosts, request_timeout=25):
-    """Fetch a URL and return its text content (or None on failure)."""
-    parsed = urllib.parse.urlparse(url)
-    host = parsed.hostname or ""
-
-    if not allowlisted_host(host, allowed_hosts):
-        return None
-
-    try:
-        req = urllib.request.Request(
-            url,
-            headers={"User-Agent": "ai-pr-reviewer/1.0"},
-        )
-        with urllib.request.urlopen(req, timeout=request_timeout) as resp:
-            raw = resp.read()
-            text = raw.decode("utf-8", errors="replace")
-            return text[:5000]
-    except Exception:
-        return None
-
-
-def _resolve_workspace_path(path, workspace_root):
-    """Resolve a workspace-relative path with traversal/symlink/sensitive guards.
-
-    Returns ``(resolved_Path, None)`` on success or ``(None, error_str)``.
-    Shared by read_file, git_log, and git_blame so they enforce the identical
-    containment + sensitive-file policy — git_blame in particular renders file
-    *content*, so it must honour the same .env/.pem/credentials blocks.
-    """
-    # Reject embedded null bytes before touching the filesystem: pathlib raises
-    # ValueError (not OSError) on them, and a NUL can truncate the path at the C
-    # layer of an underlying syscall, so an early explicit reject is safest.
-    if "\x00" in path:
-        return None, "Null byte in path"
-
-    root = Path(workspace_root).resolve()
-    try:
-        # resolve() also collapses symlinks, so a symlink that lives inside the
-        # workspace but points outside it is normalised to its real target and
-        # caught by the containment check below.
-        resolved = (root / path).resolve()
-    except (OSError, ValueError):
-        return None, f"Cannot resolve path: {path}"
-
-    # Containment via is_relative_to, NOT str.startswith: startswith wrongly
-    # accepts a sibling directory whose name shares the workspace as a prefix
-    # (e.g. resolving to /work/repo2 passes a /work/repo prefix test).
-    if not resolved.is_relative_to(root):
-        return None, "Path escapes workspace root"
-
-    if SENSITIVE_PATH_RE.search(resolved.name):
-        return None, f"Sensitive file blocked: {resolved.name}"
-
-    for deny in GH_DENY_SUBSTRINGS:
-        if deny in str(resolved):
-            return None, f"Path denied: {deny}"
-
-    return resolved, None
-
-
-def read_file(path, workspace_root, offset=None, limit=None):
-    """Read a file, optionally a 1-based line window, with path protection.
-
-    ``offset``/``limit`` let the model read a slice of a large file without
-    blowing the response cap — and cover diff-context expansion ("show me N
-    lines around this hunk") without a separate tool.
-    """
-    resolved, err = _resolve_workspace_path(path, workspace_root)
-    if err:
-        return {"error": err}
-
-    try:
-        content = resolved.read_text(encoding="utf-8", errors="replace")
-    except Exception as exc:
-        return {"error": str(exc)}
-
-    if offset is None and limit is None:
-        return {"content": content[:12000]}
-
-    lines = content.splitlines(keepends=True)
-    start = max((offset or 1) - 1, 0)
-    end = start + limit if limit is not None else len(lines)
-    window = "".join(lines[start:end])
-    return {
-        "content": window[:12000],
-        "range": {"offset": start + 1, "lines": len(lines[start:end]), "total_lines": len(lines)},
-    }
-
-
-def git_grep(pattern, workspace_root, request_timeout=15):
-    """Run git grep and return matched lines."""
-    try:
-        result = subprocess.run(
-            ["git", "grep", "-n", "--", pattern, "."],
-            cwd=workspace_root,
-            capture_output=True,
-            text=True,
-            timeout=request_timeout,
-        )
-        if result.returncode not in (0, 1):
-            return {"error": f"git grep failed: {result.stderr.strip()}"}
-        lines = result.stdout.strip().splitlines()[:60]
-        return {"matches": lines}
-    except subprocess.TimeoutExpired:
-        return {"error": f"git grep timed out after {request_timeout}s"}
-    except Exception as exc:
-        return {"error": str(exc)}
-
-
-def git_log(path, workspace_root, max_count=20, request_timeout=15):
-    """Read-only recent commit history (oneline), optionally scoped to a path.
-
-    No patch (`-p`) — subjects/metadata only, not file content. A path is run
-    as a ``-- <path>`` pathspec so it can't be read as a flag, and git keeps it
-    inside the repo regardless; the sensitive/containment guard is applied for
-    consistency with the content-bearing tools.
-    """
-    args = [
-        "git", "log", f"-n{max_count}", "--no-color",
-        "--date=short", "--pretty=format:%h %ad %an %s",
-    ]
-    if path:
-        resolved, err = _resolve_workspace_path(path, workspace_root)
-        if err:
-            return {"error": err}
-        args += ["--", str(resolved)]
-    try:
-        result = subprocess.run(
-            args, cwd=workspace_root, capture_output=True, text=True, timeout=request_timeout
-        )
-        if result.returncode != 0:
-            return {"error": f"git log failed: {result.stderr.strip()}"}
-        return {"log": result.stdout.strip().splitlines()[:max_count]}
-    except subprocess.TimeoutExpired:
-        return {"error": f"git log timed out after {request_timeout}s"}
-    except Exception as exc:
-        return {"error": str(exc)}
-
-
-def git_blame(path, workspace_root, start=None, end=None, request_timeout=15):
-    """Read-only line-level authorship for a tracked file (optional L range).
-
-    git blame renders file *content*, so the sensitive/containment guard is
-    mandatory — a committed .env/.pem must not be readable through blame.
-    """
-    resolved, err = _resolve_workspace_path(path, workspace_root)
-    if err:
-        return {"error": err}
-    args = ["git", "blame", "-w"]
-    if start is not None and end is not None:
-        args += ["-L", f"{int(start)},{int(end)}"]
-    args += ["--", str(resolved)]
-    try:
-        result = subprocess.run(
-            args, cwd=workspace_root, capture_output=True, text=True, timeout=request_timeout
-        )
-        if result.returncode != 0:
-            return {"error": f"git blame failed: {result.stderr.strip()}"}
-        return {"blame": result.stdout}
-    except subprocess.TimeoutExpired:
-        return {"error": f"git blame timed out after {request_timeout}s"}
-    except (ValueError, Exception) as exc:  # int() on a bad range → clean error
-        return {"error": str(exc)}
-
-
-def gh_api(endpoint, allowed_repos, current_repo, request_timeout=25):
-    """Make a host-platform API call with path/endpoint restrictions.
-
-    Thin shim over :func:`pr_reviewer.platform.gh_api` so the gh_api tool
-    routes through the platform seam (#226). The seam owns the allowlist
-    (path traversal, repo key, denied substrings) and the per-backend
-    transport; this shim exists only for backward compatibility with
-    call sites that import the function from this module.
-    """
-    # Imported lazily so this module can still be loaded when the
-    # platform seam is unavailable (e.g. in a script-only test that
-    # doesn't add the package to sys.path).
-    from pr_reviewer.platform import gh_api as _platform_gh_api
-    return _platform_gh_api(endpoint, allowed_repos, current_repo, request_timeout)
-
-
-def web_fetch(url, allowed_hosts, request_timeout=25):
-    """Fetch a URL using the same host-allowlist logic."""
-    parsed = urllib.parse.urlparse(url)
-    host = parsed.hostname or ""
-
-    if not allowlisted_host(host, allowed_hosts):
-        return {"error": f"Host not allowlisted: {host}"}
-
-    try:
-        req = urllib.request.Request(
-            url,
-            headers={"User-Agent": "ai-pr-reviewer/1.0"},
-        )
-        with urllib.request.urlopen(req, timeout=request_timeout) as resp:
-            raw = resp.read()
-            text = raw.decode("utf-8", errors="replace")
-            return {"content": text[:10000]}
-    except Exception as exc:
-        return {"error": str(exc)}
-
-
-def web_search(query, search_url, request_timeout=20, max_results=5):
-    """Query a configured search engine (SearXNG JSON API) for a free-text query.
-
-    ``search_url`` is the engine's search endpoint (e.g.
-    ``https://search.example.com/search``); the query and ``format=json`` are
-    appended. Returns ``{"results": [{title, url, snippet}], ...}`` capped at
-    ``max_results``, or ``{"error": ...}``. The endpoint is a single trusted,
-    operator-configured URL — unlike web_fetch it is not host-allowlisted,
-    because the model supplies only the query string, never the host.
-    """
-    if not search_url:
-        return {"error": "Search is not configured (no search_url)."}
-    sep = "&" if urllib.parse.urlparse(search_url).query else "?"
-    full = f"{search_url}{sep}" + urllib.parse.urlencode({"q": query, "format": "json"})
-    try:
-        req = urllib.request.Request(
-            full,
-            headers={"User-Agent": "ai-pr-reviewer/1.0", "Accept": "application/json"},
-        )
-        with urllib.request.urlopen(req, timeout=request_timeout) as resp:
-            data = json.loads(resp.read().decode("utf-8", errors="replace"))
-    except Exception as exc:
-        return {"error": str(exc)}
-
-    results = []
-    for item in (data.get("results") or [])[:max_results]:
-        if not isinstance(item, dict):
-            continue
-        results.append({
-            "title": str(item.get("title", ""))[:300],
-            "url": str(item.get("url", "")),
-            "snippet": str(item.get("content", ""))[:500],
-        })
-    return {"results": results}
-
-
-def run_command(command, workspace_root, request_timeout=30):
-    """Execute a named read-only command definition.
-
-    The planner may choose only command names from ALLOWED_COMMANDS. Raw shell
-    text is intentionally rejected so untrusted PR/corpus content cannot shape
-    a bash command line.
-    """
-    command_name = (command or "").strip()
-    args = ALLOWED_COMMANDS.get(command_name)
-    if args is None:
-        return {
-            "error": (
-                "Command not allowlisted. Use one of: "
-                + command_catalog_markdown()
-            )
-        }
-
-    try:
-        result = subprocess.run(
-            args,
-            cwd=workspace_root,
-            capture_output=True,
-            text=True,
-            timeout=request_timeout,
-        )
-        return {
-            "stdout": mask_secrets((result.stdout or "").strip()),
-            "stderr": mask_secrets((result.stderr or "").strip()),
-            "exit_code": result.returncode,
-            "command": command_name,
-        }
-    except subprocess.TimeoutExpired as exc:
-        stdout = exc.stdout or ""
-        stderr = exc.stderr or ""
-        if isinstance(stdout, bytes):
-            stdout = stdout.decode("utf-8", errors="replace")
-        if isinstance(stderr, bytes):
-            stderr = stderr.decode("utf-8", errors="replace")
-        return {
-            "error": f"Command timed out after {request_timeout}s",
-            "stdout": mask_secrets(stdout),
-            "stderr": mask_secrets(stderr),
-            "command": command_name,
-        }
 
 
 def build_planning_context(max_bytes, corpus_path=None):
@@ -778,183 +312,6 @@ def build_results_feedback(executed, max_bytes):
         )
     text, _truncated = mask_and_truncate("\n".join(lines), max_bytes)
     return text
-
-
-def execute_tool_request(
-    tool_name,
-    args,
-    workspace_root,
-    allowed_gh_repos,
-    current_repo,
-    allowed_hosts,
-    max_response_bytes,
-    request_timeout,
-    search_url="",
-    max_search_results=5,
-):
-    """Execute a single tool request and return the result dict.
-
-    Shared by both file-based and direct planning paths to avoid duplication
-    of validation, execution, truncation, and error-handling logic.
-    """
-    tool_result = {"tool": tool_name, "status": "error", "result": {}}
-
-    try:
-        if tool_name == "read_file":
-            path = args.get("path", "")
-            if not path:
-                raise ValueError("Missing 'path' argument")
-            res = read_file(
-                path, workspace_root, _opt_int(args.get("offset")), _opt_int(args.get("limit"))
-            )
-            if res.get("error"):
-                raise ValueError(res["error"])
-            text = mask_secrets(res.get("content", ""))
-            text, _ = mask_and_truncate(text, max_response_bytes)
-            result_payload = {"content": text}
-            if res.get("range"):
-                result_payload["range"] = res["range"]
-            tool_result["result"] = result_payload
-
-        elif tool_name == "git_log":
-            max_count = max(1, min(_opt_int(args.get("max_count")) or 20, 100))
-            res = git_log(
-                args.get("path", "") or "", workspace_root, max_count, request_timeout
-            )
-            if res.get("error"):
-                raise ValueError(res["error"])
-            text, _ = mask_and_truncate("\n".join(res.get("log", [])), max_response_bytes)
-            tool_result["result"] = {"log": text}
-
-        elif tool_name == "git_blame":
-            path = args.get("path", "")
-            if not path:
-                raise ValueError("Missing 'path' argument")
-            res = git_blame(
-                path, workspace_root,
-                _opt_int(args.get("start")), _opt_int(args.get("end")), request_timeout,
-            )
-            if res.get("error"):
-                raise ValueError(res["error"])
-            text, _ = mask_and_truncate(res.get("blame", ""), max_response_bytes)
-            tool_result["result"] = {"blame": text}
-
-        elif tool_name == "git_grep":
-            pattern = args.get("pattern", "")
-            if not pattern:
-                raise ValueError("Missing 'pattern' argument")
-            res = git_grep(pattern, workspace_root, request_timeout)
-            if res.get("error"):
-                raise ValueError(res["error"])
-            matches = res.get("matches", [])
-            text = "\n".join(matches)
-            text, _ = mask_and_truncate(text, max_response_bytes)
-            tool_result["result"] = {"matches": matches[:60]}
-
-        elif tool_name == "gh_api":
-            endpoint = args.get("endpoint", "")
-            if not endpoint:
-                raise ValueError("Missing 'endpoint' argument")
-            res = gh_api(endpoint, allowed_gh_repos, current_repo, request_timeout)
-            if res.get("error"):
-                raise ValueError(res["error"])
-            data = res.get("data")
-            text = ""
-            if isinstance(data, (dict, list)):
-                text = json.dumps(data, indent=2)[:max_response_bytes]
-            tool_result["result"] = {"response": text}
-
-        elif tool_name == "web_fetch":
-            url = args.get("url", "")
-            if not url:
-                raise ValueError("Missing 'url' argument")
-            res = web_fetch(url, allowed_hosts, request_timeout)
-            if res.get("error"):
-                raise ValueError(res["error"])
-            content_text = res.get("content", "")
-            text, _ = mask_and_truncate(content_text, max_response_bytes)
-            tool_result["result"] = {"content": text}
-
-        elif tool_name == "web_search":
-            query = args.get("query", "")
-            if not query:
-                raise ValueError("Missing 'query' argument")
-            res = web_search(query, search_url, request_timeout, max_search_results)
-            if res.get("error"):
-                raise ValueError(res["error"])
-            text = json.dumps(res.get("results", []), indent=2)
-            text, _ = mask_and_truncate(text, max_response_bytes)
-            tool_result["result"] = {"results": text}
-
-        elif tool_name == "run_command":
-            command = args.get("command", "")
-            if not command:
-                raise ValueError("Missing 'command' argument")
-            res = run_command(command, workspace_root, request_timeout)
-            if res.get("error"):
-                raise ValueError(res["error"])
-            stdout_text = res.get("stdout", "")
-            stderr_text = res.get("stderr", "")
-            stdout_text, _ = mask_and_truncate(stdout_text, max_response_bytes)
-            stderr_text, _ = mask_and_truncate(stderr_text, max_response_bytes)
-            tool_result["result"] = {
-                "stdout": stdout_text,
-                "stderr": stderr_text,
-                "exit_code": res.get("exit_code"),
-                "command": res.get("command"),
-            }
-
-        else:
-            raise ValueError(f"Unknown tool: {tool_name}")
-
-        tool_result["status"] = "ok"
-    except Exception as exc:
-        # Error messages from raised ValueError (from res["error"] checks
-        # above) are masked by mask_secrets() in write_outputs(), which
-        # processes the markdown output. This is consistent with how
-        # run_command error messages are redacted.
-        tool_result["result"] = {"error": str(exc)}
-
-    return tool_result
-
-
-def execute_tool_requests(
-    normalized_requests,
-    workspace_root,
-    allowed_gh_repos,
-    current_repo,
-    allowed_hosts,
-    max_response_bytes,
-    request_timeout,
-):
-    """Execute normalized (tool_name, args) requests concurrently.
-
-    Tools are read-only, so they are safe to run in parallel; results are
-    returned in request order so the markdown/JSON output stays deterministic.
-    """
-    if not normalized_requests:
-        return []
-
-    def _run(pair):
-        tool_name, args = pair
-        return execute_tool_request(
-            tool_name,
-            args,
-            workspace_root,
-            allowed_gh_repos,
-            current_repo,
-            allowed_hosts,
-            max_response_bytes,
-            request_timeout,
-        )
-
-    if len(normalized_requests) == 1:
-        return [_run(normalized_requests[0])]
-
-    with ThreadPoolExecutor(
-        max_workers=min(4, len(normalized_requests))
-    ) as executor:
-        return list(executor.map(_run, normalized_requests))
 
 
 def tool_result_md_lines(index, tool_name, args, tool_result):
